% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProteinStucture.R
\name{getProteinStucture}
\alias{getProteinStucture}
\title{function to parse the pdb file from PDB database}
\usage{
getProteinStucture(pdb)
}
\arguments{
\item{pdb}{a pdbfile}
}
\value{
list
}
\description{
function to parse the pdb file from PDB database
}
\examples{
getProteinStucture("3o8o.pdb")


}
