% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_function.R
\name{idmapping}
\alias{idmapping}
\title{Use r code to connect uniprot web api}
\usage{
idmapping(query, inputid, outputid, fmt)
}
\arguments{
\item{query}{vector of protein ids}

\item{inputid}{type of input id, characterr}

\item{outputid}{type of output id, character}

\item{fmt}{output form}
}
\value{
a dataframe
}
\description{
You can choose one id as input id as well as another id at output id
know more information you can learnf rom uniprot documatation
}
\examples{
result <- idmapping(query = proid, inputid ='ACC', outputid = "P_ENTREZGENEID", fmt = "fmt")
}
