% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_function.R
\name{print_R_package}
\alias{print_R_package}
\title{Title
R package management
print all installed R packages and its version
print all installed R packages and its version
print all installed R packages and its version}
\usage{
print_R_package(output = "Installed_R_package.txt")
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Title
R package management
print all installed R packages and its version
print all installed R packages and its version
print all installed R packages and its version
}
\examples{
print_R_package()
}
