% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitAndCombine.R
\name{splitAndCombine}
\alias{splitAndCombine}
\title{fast estabolish the mapping relation between two element}
\usage{
splitAndCombine(gene, rxn, sep0)
}
\arguments{
\item{gene}{vetor}

\item{rxn}{vector}

\item{sep0}{string}
}
\value{
a dataframe
}
\description{
fast estabolish the mapping relation between two element
}
\examples{
gene <- c('a&b','c')
rxn <- c('r1','r2')
splitAndCombine(gene, rxn, sep0="&")

}
