% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_function.R
\name{AutoUpdate}
\alias{AutoUpdate}
\title{Title
using the description1 in para1 to update the description2 in para2
these parameters should not be factor level}
\usage{
AutoUpdate(description1, para1, description2, para2)
}
\arguments{
\item{para2}{}
}
\value{

}
\description{
Title
using the description1 in para1 to update the description2 in para2
these parameters should not be factor level
}
