% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingleReactionFormula.R
\name{getSingleReactionFormula}
\alias{getSingleReactionFormula}
\title{get single mapping result for one element}
\usage{
getSingleReactionFormula(description, reaction, ko)
}
\arguments{
\item{description}{vector contains the description for each element from 'reaction'}

\item{reaction}{vector contrains the element been mapped}

\item{ko}{vector contains the element need mapping}
}
\value{
one element
}
\description{
get single mapping result for one element
}
